requirements.txt
pandas==1.5.3
plotly==5.13.1
python_dateutil==2.8.2
streamlit==1.19.0
